backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.7.1
PyMySQL==1.1.1
requests==2.32.3
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
python-dotenv==1.0.1
werkzeug==3.1.3
pytest
pytest-mock
